requirements.txt
pandas>=2.1.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
sqlalchemy>=2.0.0
mysql-connector-python>=8.3.0
pymysql>=1.1.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0
tqdm>=4.66.0
streamlit>=1.32.0
plotly>=5.18.0
